requirements.txt
alembic
annotated-types
blinker
Brotli
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
cryptography
firebase-admin
Flask
Flask-Compress
Flask-Cors
Flask-Migrate
Flask-SQLAlchemy
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
httplib2
idna
itsdangerous
Jinja2
Mako
MarkupSafe
msgpack
proto-plus
protobuf
psycopg2
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
PyJWT
pyparsing
python-dotenv
requests
rsa
SQLAlchemy
tqdm
typing_extensions
uritemplate
urllib3
Werkzeug
